refactor: user 패키지 이동
